+++ ikea-zigbee-drivers/src/capabilities/Brightness.groovy
@@ -1,5 +1,6 @@
 {{!--------------------------------------------------------------------------}}
 {{# @definition }}
+capability 'Light'
 capability 'ChangeLevel'
 capability 'SwitchLevel'
 {{/ @definition }}